NAV-88 Validate graph after loading.
