+++ navigator_engine/tests/integration_tests/test_end_to_end.py
@@ -23,50 +23,3 @@ def test_end_to_end(client):
     }))
     assert response.status_code == 200
     assert response.json['decision']['id'] == 'EST-OVV-02-A'
-
-
-def validate_pluggable_logic(node, network):
-
-    function_string = node.conditional.function
-    function_name, function_args = common.get_pluggable_function_and_args(function_string)
-
-    if function_name.startswith('check_manual_confirmation'):
-        arg_action_is_child = False
-        for node, child_node in network.out_edges(node):
-            if child_node.ref == function_args[0]:
-                arg_action_is_child = True
-        assert arg_action_is_child, f"{node.ref} has bad function {function_string}"
-
-
-def validate_graph(graph_id):
-    graph = model.load_graph(graph_id)
-    network = graph.to_networkx()
-    milestones = []
-
-    complete_node = None
-    for node, out_degree in network.out_degree():
-
-        if getattr(node, 'action_id'):
-            assert out_degree == 0, f"{node.ref} has wrong out_degree"
-            assert node.action.title, f"{node.ref} has no title specified"
-            assert node.action.html, f"{node.ref} has no content specified"
-        elif getattr(node, 'conditional_id'):
-            assert out_degree == 2, f"{node.ref} has wrong out_degree"
-            assert node.conditional.function, f"{node.ref} has no test function"
-            edge_types = [edge[2] for edge in network.out_edges([node], 'type')]
-            assert set(edge_types) == {True, False}, f"{node.ref} has wrong out edges"
-            validate_pluggable_logic(node, network)
-        elif getattr(node, 'milestone_id'):
-            assert out_degree == 1, f"{node.ref} has wrong out_degree"
-            assert node.milestone.graph_id, f"{node.ref} has no graph ID"
-            edge_types = [edge[2] for edge in network.out_edges([node], 'type')]
-            assert edge_types == [True], f"{node.ref} has wrong out edges"
-            milestones.append(node.milestone.graph_id)
-
-        if getattr(node, 'action') and node.action.complete:
-            complete_node = node
-
-    assert complete_node, "Graph {graph_id} has no complete node"
-
-    for milestone in milestones:
-        validate_graph(milestone)
